+++ prismaDmob/sql/getClientProviderDistribution.sql
@@ -8,10 +8,10 @@ with "miner_pieces" as (select 'f0' || "clientId"       as "client",
                         where "termStart" >= 3847920                                           -- nv22 start
                           and to_timestamp("termStart" * 30 + 1598306400) <= current_timestamp -- deals that didn't start yet
                         group by "client", "provider", "pieceCid")
-select "client"                      as "client",
-       "provider"                    as "provider",
-       sum("total_deal_size")::float as "total_deal_size",
-       sum("piece_size")::float      as "unique_data_size",
-       sum("claims_count")::float    as "claims_count"
+select "client"                       as "client",
+       "provider"                     as "provider",
+       sum("total_deal_size")::bigint as "total_deal_size",
+       sum("piece_size")::bigint      as "unique_data_size",
+       sum("claims_count")::bigint    as "claims_count"
 from "miner_pieces"
 group by "client", "provider";
